Migrated reprocess tests to be in autonicer tests
